backend/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
requests==2.31.0
Werkzeug==3.0.1
python-dotenv==1.0.0

